+++ src/get/decompress.js
@@ -1,34 +1,35 @@
-import path from "node:path";
+import { resolve } from "node:path";
 
-import extract from "extract-zip";
-import tar from "tar";
+import Decompress from "decompress";
 
-const decompress = (platform, outDir) => {
-  return new Promise((resolve, reject) => {
-    if (platform === "linux") {
-      tar
-        .x({
-          file: `${outDir}/nw.tar.gz`,
-          C: `${outDir}`,
-        })
-        .then(() => {
-          resolve();
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    } else {
-      extract(path.resolve(`${outDir}/nw.zip`), {
-        dir: path.resolve(`${outDir}`),
-      })
-        .then(() => {
-          resolve();
-        })
-        .catch((error) => {
-          reject(error);
-        });
+import { log } from "../log.js";
+
+/**
+ * Decompress NW.js binary
+ *
+ * @param  {string}        platform     Platform
+ * @param  {string}        cacheDir     Output directory
+ * @param  {string}        downloadUrl  Download url
+ * @return {Promise<void>}
+ */
+const decompress = async (platform, cacheDir, downloadUrl) => {
+  try {
+    if (downloadUrl === "https://dl.nwjs.io") {
+      if (platform === "linux") {
+        await Decompress(resolve(cacheDir, "nw.tar.gz"), cacheDir);
+      } else {
+        await Decompress(resolve(cacheDir, "nw.zip"), cacheDir);
+      }
+    } else if (
+      downloadUrl ===
+      "https://github.com/nwjs-ffmpeg-prebuilt/nwjs-ffmpeg-prebuilt/releases/download"
+    ) {
+      await Decompress(resolve(cacheDir, "ffmpeg.zip"), cacheDir);
     }
-  });
+  } catch (error) {
+    log.error(error);
+    throw error;
+  }
 };
 
 export { decompress };