+++ src/get/download.js
@@ -1,53 +1,84 @@
 import fs from "node:fs";
-import https from "node:https";
+import { resolve } from "node:path";
+import https from "https";
 
 import progress from "cli-progress";
 
 const bar = new progress.SingleBar({}, progress.Presets.rect);
 
+/**
+ * Download NW.js binary
+ *
+ * @param  {string}        version       Version
+ * @param  {string}        flavor        Flavor
+ * @param  {string}        platform      Platform
+ * @param  {string}        architecture  Architecture
+ * @param  {string}        downloadUrl   Download url
+ * @param  {string}        cacheDir      Output directory
+ * @return {Promise<void>}
+ */
 const download = (
   version,
   flavor,
   platform,
   architecture,
   downloadUrl,
-  outDir,
+  cacheDir,
 ) => {
-  return new Promise((resolve, reject) => {
-    if (downloadUrl !== "https://dl.nwjs.io") {
-      reject(new Error("Invalid download url. Please try again."));
+  let url;
+  let out;
+  return new Promise((res, rej) => {
+    if (downloadUrl === "https://dl.nwjs.io") {
+      url = `${downloadUrl}/v${version}/nwjs${
+        flavor === "sdk" ? "-sdk" : ""
+      }-v${version}-${platform}-${architecture}.${
+        platform === "linux" ? "tar.gz" : "zip"
+      }`;
+      out = resolve(cacheDir, `nw.${platform === "linux" ? "tar.gz" : "zip"}`);
+    } else if (
+      downloadUrl ===
+      "https://github.com/nwjs-ffmpeg-prebuilt/nwjs-ffmpeg-prebuilt/releases/download"
+    ) {
+      url = `${downloadUrl}/${version}/${version}-${platform}-${architecture}.zip`;
+      out = resolve(cacheDir, `ffmpeg.zip`);
+    } else {
+      rej(new Error("Invalid download url. Please try again."));
     }
 
-    let url = `${downloadUrl}/v${version}/nwjs${
-      flavor === "sdk" ? "-sdk" : ""
-    }-v${version}-${platform}-${architecture}.${
-      platform === "linux" ? "tar.gz" : "zip"
-    }`;
+    https.get(url, (response) => {
+      if (
+        downloadUrl ===
+        "https://github.com/nwjs-ffmpeg-prebuilt/nwjs-ffmpeg-prebuilt/releases/download"
+      ) {
+        url = response.headers.location;
+      }
 
-    https.get(url, (res) => {
-      let chunks = 0;
-      bar.start(Number(res.headers["content-length"]), 0);
+      https.get(url, (response) => {
+        let chunks = 0;
+        bar.start(Number(response.headers["content-length"]), 0);
+        response.on("data", (chunk) => {
+          chunks += chunk.length;
+          bar.increment();
+          bar.update(chunks);
+        });
 
-      res.on("data", (chunk) => {
-        chunks += chunk.length;
-        bar.increment();
-        bar.update(chunks);
+        response.on("error", (error) => {
+          rej(error);
+        });
+
+        response.on("end", () => {
+          bar.stop();
+          res();
+        });
+
+        fs.mkdirSync(cacheDir, { recursive: true });
+        const stream = fs.createWriteStream(out);
+        response.pipe(stream);
       });
 
-      res.on("error", (error) => {
-        reject(error);
+      response.on("error", (error) => {
+        rej(error);
       });
-
-      res.on("end", () => {
-        bar.stop();
-        resolve();
-      });
-
-      fs.mkdirSync(outDir, { recursive: true });
-      const stream = fs.createWriteStream(
-        `${outDir}/nw.${platform === "linux" ? "tar.gz" : "zip"}`,
-      );
-      res.pipe(stream);
     });
   });
 };